refactor: Teachers Section in own file
